Fix purge position

+++ programs/helium-sub-daos/src/instructions/staking/stake_v0.rs
@@ -85,7 +85,7 @@ pub fn handler(ctx: Context<StakeV0>) -> Result<()> {
 
   // update the stake
   update_subdao_vehnt(sub_dao, curr_ts);
-  sub_dao.vehnt_staked = sub_dao.vehnt_staked.checked_add(available_vehnt).unwrap();
+  sub_dao.vehnt_staked = sub_dao.vehnt_staked.checked_add(i128::from(available_vehnt)).unwrap();
   sub_dao.vehnt_fall_rate = sub_dao
     .vehnt_fall_rate
     .checked_sub(stake_position.fall_rate)
