+++ programs/helium-sub-daos/src/state.rs
@@ -101,7 +101,7 @@ pub struct SubDaoV0 {
   pub treasury: Pubkey,       // Treasury of HNT
   pub rewards_escrow: Pubkey, // Escrow account for DNT rewards
   pub staker_pool: Pubkey,    // Pool of DNT tokens which veHNT stakers can claim from
-  pub vehnt_staked: u64,
+  pub vehnt_staked: i128, // This can go negative if a position isn't purged in time and so the fall rate is over-applied. Then corrected in purge_position_v0
   pub vehnt_last_calculated_ts: i64,
   pub vehnt_fall_rate: u128, // the vehnt amount that the position decays by per second, with 12 decimals of extra precision
   pub authority: Pubkey,
